+++ example/applauncher/applauncher.js
@@ -94,7 +94,9 @@ const Applauncher = ({ width = 500, height = 500, spacing = 12 }) => {
 // there needs to be only one instance
 export const applauncher = Widget.Window({
     name: WINDOW_NAME,
-    popup: true,
+    setup: self => self.keybind('Escape', () => {
+        App.closeWindow(WINDOW_NAME)
+    }),
     visible: false,
     keymode: 'exclusive',
     child: Applauncher({