+++ example/notification-popups/config.js
@@ -1,19 +1,18 @@
-import App from 'resource:///com/github/Aylur/ags/app.js';
 import { notificationPopup } from './notificationPopups.js';
-import { execAsync, timeout } from 'resource:///com/github/Aylur/ags/utils.js';
 
-timeout(100, () => execAsync([
-    'notify-send',
-    'Notification Popup Example',
-    'Lorem ipsum dolor sit amet, qui minim labore adipisicing ' +
-    'minim sint cillum sint consectetur cupidatat.',
-    '-A', 'Cool!',
-    '-i', 'info-symbolic',
-]));
+Utils.timeout(100, () => Utils.notify({
+    summary: 'Notification Popup Example',
+    iconName: "info-symbolic",
+    body: 'Lorem ipsum dolor sit amet, qui minim labore adipisicing '
+        + 'minim sint cillum sint consectetur cupidatat.',
+    actions: {
+        "Cool": () => print("pressed Cool"),
+    },
+}))
 
-export default {
+App.config({
     style: App.configDir + '/style.css',
     windows: [
         notificationPopup,
     ],
-};
+});